start refactoring names and structs
