+++ wgctrl-rs/src/device.rs
@@ -1,11 +1,11 @@
 use libc::c_char;
 
-use crate::{backends, key::Key};
+use crate::{backends, key::Key, KeyPair, PeerConfigBuilder};
 
 use std::{
     borrow::Cow,
     ffi::CStr,
-    fmt,
+    fmt, io,
     net::{IpAddr, SocketAddr},
     str::FromStr,
     time::SystemTime,
@@ -86,7 +86,7 @@ pub struct PeerInfo {
 /// The peer statistics are retrieved once at construction time,
 /// and need to be updated manually by calling [`get_by_name`](DeviceInfo::get_by_name).
 #[derive(Debug, PartialEq, Eq, Clone)]
-pub struct DeviceInfo {
+pub struct Device {
     /// The interface name of this device
     pub name: InterfaceName,
     /// The public encryption key of this interface (if present)
@@ -223,7 +223,7 @@ impl From<InvalidInterfaceName> for std::io::Error {
 
 impl std::error::Error for InvalidInterfaceName {}
 
-impl DeviceInfo {
+impl Device {
     /// Enumerates all WireGuard interfaces currently present in the system
     /// and returns their names.
     ///
@@ -231,11 +231,10 @@ impl DeviceInfo {
     /// detailed information on each interface.
     #[cfg(target_os = "linux")]
     pub fn enumerate() -> Result<Vec<InterfaceName>, std::io::Error> {
-        if backends::kernel::exists() {
-            backends::kernel::enumerate()
-        } else {
-            backends::userspace::enumerate()
-        }
+        backends::kernel::enumerate()
+        // } else {
+        //     backends::userspace::enumerate()
+        // }
     }
 
     #[cfg(not(target_os = "linux"))]
@@ -245,12 +244,7 @@ impl DeviceInfo {
 
     #[cfg(target_os = "linux")]
     pub fn get_by_name(name: &InterfaceName) -> Result<Self, std::io::Error> {
-        if backends::kernel::exists() {
-            backends::kernel::get_by_name(name)
-        } else {
-            println!("kernel module not detected. falling back to userspace backend.");
-            backends::userspace::get_by_name(name)
-        }
+        backends::kernel::get_by_name(name).or_else(|_| backends::userspace::get_by_name(name))
     }
 
     #[cfg(not(target_os = "linux"))]
@@ -269,11 +263,190 @@ impl DeviceInfo {
     }
 }
 
+/// Builds and represents a configuration that can be applied to a WireGuard interface.
+///
+/// This is the primary way of changing the settings of an interface.
+///
+/// Note that if an interface exists, the configuration is applied _on top_ of the existing
+/// settings, and missing parts are not overwritten or set to defaults.
+///
+/// If this is not what you want, use [`delete_interface`](delete_interface)
+/// to remove the interface entirely before applying the new configuration.
+///
+/// # Example
+/// ```rust
+/// # use wgctrl::*;
+/// # use std::net::AddrParseError;
+/// # fn try_main() -> Result<(), AddrParseError> {
+/// let our_keypair = KeyPair::generate();
+/// let peer_keypair = KeyPair::generate();
+/// let server_addr = "192.168.1.1:51820".parse()?;
+///
+/// DeviceConfigBuilder::new()
+///     .set_keypair(our_keypair)
+///     .replace_peers()
+///     .add_peer_with(&peer_keypair.public, |peer| {
+///         peer.set_endpoint(server_addr)
+///             .replace_allowed_ips()
+///             .allow_all_ips()
+///     }).apply(&"wg-example".parse().unwrap());
+///
+/// println!("Send these keys to your peer: {:#?}", peer_keypair);
+///
+/// # Ok(())
+/// # }
+/// # fn main() { try_main(); }
+/// ```
+#[derive(Debug, PartialEq, Eq, Clone)]
+pub struct DeviceUpdate {
+    pub(crate) public_key: Option<Key>,
+    pub(crate) private_key: Option<Key>,
+    pub(crate) fwmark: Option<u32>,
+    pub(crate) listen_port: Option<u16>,
+    pub(crate) peers: Vec<PeerConfigBuilder>,
+    pub(crate) replace_peers: bool,
+}
+
+impl DeviceUpdate {
+    /// Creates a new `DeviceConfigBuilder` that does nothing when applied.
+    pub fn new() -> Self {
+        DeviceUpdate {
+            public_key: None,
+            private_key: None,
+            fwmark: None,
+            listen_port: None,
+            peers: vec![],
+            replace_peers: false,
+        }
+    }
+
+    /// Sets a new keypair to be applied to the interface.
+    ///
+    /// This is a convenience method that simply wraps
+    /// [`set_public_key`](DeviceConfigBuilder::set_public_key)
+    /// and [`set_private_key`](DeviceConfigBuilder::set_private_key).
+    pub fn set_keypair(self, keypair: KeyPair) -> Self {
+        self.set_public_key(keypair.public)
+            .set_private_key(keypair.private)
+    }
+
+    /// Specifies a new public key to be applied to the interface.
+    pub fn set_public_key(mut self, key: Key) -> Self {
+        self.public_key = Some(key);
+        self
+    }
+
+    /// Specifies that the public key for this interface should be unset.
+    pub fn unset_public_key(self) -> Self {
+        self.set_public_key(Key::zero())
+    }
+
+    /// Sets a new private key to be applied to the interface.
+    pub fn set_private_key(mut self, key: Key) -> Self {
+        self.private_key = Some(key);
+        self
+    }
+
+    /// Specifies that the private key for this interface should be unset.
+    pub fn unset_private_key(self) -> Self {
+        self.set_private_key(Key::zero())
+    }
+
+    /// Specifies the fwmark value that should be applied to packets coming from the interface.
+    pub fn set_fwmark(mut self, fwmark: u32) -> Self {
+        self.fwmark = Some(fwmark);
+        self
+    }
+
+    /// Specifies that fwmark should not be set on packets from the interface.
+    pub fn unset_fwmark(self) -> Self {
+        self.set_fwmark(0)
+    }
+
+    /// Specifies the port to listen for incoming packets on.
+    ///
+    /// This is useful for a server configuration that listens on a fixed endpoint.
+    pub fn set_listen_port(mut self, port: u16) -> Self {
+        self.listen_port = Some(port);
+        self
+    }
+
+    /// Specifies that a random port should be used for incoming packets.
+    ///
+    /// This is probably what you want in client configurations.
+    pub fn randomize_listen_port(self) -> Self {
+        self.set_listen_port(0)
+    }
+
+    /// Specifies a new peer configuration to be added to the interface.
+    ///
+    /// See [`PeerConfigBuilder`](PeerConfigBuilder) for details on building
+    /// peer configurations. This method can be called more than once, and all
+    /// peers will be added to the configuration.
+    pub fn add_peer(mut self, peer: PeerConfigBuilder) -> Self {
+        self.peers.push(peer);
+        self
+    }
+
+    /// Specifies a new peer configuration using a builder function.
+    ///
+    /// This is simply a convenience method to make adding peers more fluent.
+    /// This method can be called more than once, and all peers will be added
+    /// to the configuration.
+    pub fn add_peer_with(
+        self,
+        pubkey: &Key,
+        builder: impl Fn(PeerConfigBuilder) -> PeerConfigBuilder,
+    ) -> Self {
+        self.add_peer(builder(PeerConfigBuilder::new(pubkey)))
+    }
+
+    /// Specifies multiple peer configurations to be added to the interface.
+    pub fn add_peers(mut self, peers: &[PeerConfigBuilder]) -> Self {
+        self.peers.extend_from_slice(peers);
+        self
+    }
+
+    /// Specifies that the peer configurations in this `DeviceConfigBuilder` should
+    /// replace the existing configurations on the interface, not modify or append to them.
+    pub fn replace_peers(mut self) -> Self {
+        self.replace_peers = true;
+        self
+    }
+
+    /// Specifies that the peer with this public key should be removed from the interface.
+    pub fn remove_peer_by_key(self, public_key: &Key) -> Self {
+        let mut peer = PeerConfigBuilder::new(public_key);
+        peer.remove_me = true;
+        self.add_peer(peer)
+    }
+
+    /// Build and apply the configuration to a WireGuard interface by name.
+    ///
+    /// An interface with the provided name will be created if one does not exist already.
+    #[cfg(target_os = "linux")]
+    pub fn apply(self, iface: &InterfaceName) -> io::Result<()> {
+        backends::kernel::apply(self, &iface)
+        // } else {
+        //     backends::userspace::apply(self, iface)
+        // }
+    }
+
+    #[cfg(not(target_os = "linux"))]
+    pub fn apply(self, iface: &InterfaceName) -> io::Result<()> {
+        backends::userspace::apply(self, iface)
+    }
+}
+
+impl Default for DeviceUpdate {
+    fn default() -> Self {
+        Self::new()
+    }
+}
+
 #[cfg(test)]
 mod tests {
-    use crate::{
-        DeviceConfigBuilder, InterfaceName, InvalidInterfaceName, KeyPair, PeerConfigBuilder,
-    };
+    use crate::{DeviceUpdate, InterfaceName, InvalidInterfaceName, KeyPair, PeerConfigBuilder};
 
     const TEST_INTERFACE: &str = "wgctrl-test";
     use super::*;
@@ -285,14 +458,14 @@ mod tests {
         }
 
         let keypairs: Vec<_> = (0..10).map(|_| KeyPair::generate()).collect();
-        let mut builder = DeviceConfigBuilder::new();
+        let mut builder = DeviceUpdate::new();
         for keypair in &keypairs {
             builder = builder.add_peer(PeerConfigBuilder::new(&keypair.public))
         }
         let interface = TEST_INTERFACE.parse().unwrap();
         builder.apply(&interface).unwrap();
 
-        let device = DeviceInfo::get_by_name(&interface).unwrap();
+        let device = Device::get_by_name(&interface).unwrap();
 
         for keypair in &keypairs {
             assert!(device
